requirements.txt
pycuda
six==1.12.0
scipy
scikit-learn==0.21.3
numpy
tensorflow-gpu==1.14.0
tensorflow-estimator==1.14.0
tensorrt==6.0.0.1